moved TestGeenericBroadcastReceiver to test package
