+++ android-sdk/src/androidTest/AndroidManifest.xml
@@ -5,14 +5,14 @@
 
     <application
         android:name="com.sensorberg.sdk.SensorbergTestApplication">
-        <receiver android:name="com.sensorberg.sdk.internal.TestGenericBroadcastReceiver"
+        <receiver android:name="com.sensorberg.sdk.test.TestGenericBroadcastReceiver"
             android:enabled="true">
             <intent-filter>
                 <action android:name="com.sensorberg.android.PRESENT_ACTION" />
             </intent-filter>
         </receiver>
 
-        <receiver android:name="com.sensorberg.sdk.internal.TestGenericBroadcastReceiver2"
+        <receiver android:name="com.sensorberg.sdk.test.TestGenericBroadcastReceiver2"
             android:process=".sensorberg"
             android:enabled="true">
             <intent-filter>
