+++ android-sdk/src/androidTest/java/com/sensorberg/sdk/internal/TheIntentSchedulingBeUpdateable.java
@@ -3,6 +3,7 @@ package com.sensorberg.sdk.internal;
 import com.sensorberg.sdk.Constants;
 import com.sensorberg.sdk.SensorbergTestApplication;
 import com.sensorberg.sdk.di.TestComponent;
+import com.sensorberg.sdk.test.TestGenericBroadcastReceiver;
 import com.sensorberg.sdk.testUtils.TestClock;
 import com.sensorberg.sdk.testUtils.TestServiceScheduler;
